+++ aoe2/aoe2_game_overlay.py
@@ -10,13 +10,12 @@ from PyQt5.QtCore import Qt, QSize
 
 from common.label_display import QLabelSettings
 from common.useful_tools import cut_name_length, widget_x_end, widget_y_end, popup_message
-from common.rts_overlay import RTSGameMatchDataOverlay, scale_list_int
+from common.rts_overlay import RTSGameOverlay, scale_list_int
 from common.build_order_tools import get_total_on_resource, get_build_orders
 from common.build_order_window import BuildOrderWindow
 
 from aoe2.aoe2_settings import AoE2OverlaySettings
 from aoe2.aoe2_build_order import check_valid_aoe2_build_order, build_order_sorting
-from aoe2.aoe2_request import get_match_data_threading, is_valid_fetch_match_data
 from aoe2.aoe2_civ_icon import aoe2_civilization_icon
 
 
@@ -24,10 +23,9 @@ from aoe2.aoe2_civ_icon import aoe2_civilization_icon
 class PanelID(Enum):
     CONFIG = 0  # Configuration
     BUILD_ORDER = 1  # Display Build Order
-    MATCH_DATA = 2  # Display Match Data
 
 
-class AoE2GameOverlay(RTSGameMatchDataOverlay):
+class AoE2GameOverlay(RTSGameOverlay):
     """Game overlay application for AoE2"""
 
     def __init__(self, app: QApplication, directory_main: str):
@@ -78,14 +76,6 @@ class AoE2GameOverlay(RTSGameMatchDataOverlay):
         self.civilization_select.setFont(QFont(layout.font_police, layout.font_size))
         self.civilization_select.adjustSize()
 
-        # match data
-        self.match_data_thread_started = False  # True after the first call to 'get_match_data_threading'
-        self.store_match_data = []  # used for url requests in parallel thread
-        self.match_data = None  # match data to use
-        self.match_data_warnings = []  # warnings related to match data not found
-        self.match_data_thread_id = None
-        self.match_data_stop_flag = Event()
-
         # initialize build orders if folder does not exist and copy the samples
         self.sample_directory_build_orders = os.path.join(self.directory_main, 'build_orders', self.name_game)
         if not os.path.isdir(self.directory_build_orders):
@@ -132,10 +122,6 @@ class AoE2GameOverlay(RTSGameMatchDataOverlay):
         self.civilization_select.setFont(QFont(layout.font_police, layout.font_size))
         self.civilization_select.adjustSize()
 
-        # game match data
-        self.match_data = None  # match data to use
-        self.match_data_warnings = []  # warnings related to match data not found
-
         # sort build orders
         self.build_orders.sort(key=build_order_sorting)
 
@@ -158,10 +144,6 @@ class AoE2GameOverlay(RTSGameMatchDataOverlay):
     def quit_application(self):
         """Quit the application"""
         super().quit_application()
-
-        self.match_data_stop_flag.set()
-        if self.match_data_thread_id is not None:
-            self.match_data_thread_id.join()
 
         self.close()
         QApplication.quit()
@@ -207,8 +189,6 @@ class AoE2GameOverlay(RTSGameMatchDataOverlay):
             self.build_order_search.setFocus()
         elif self.selected_panel == PanelID.BUILD_ORDER:  # Build Order
             self.update_build_order()
-        elif self.selected_panel == PanelID.MATCH_DATA:  # Display Match Data
-            self.update_match_data_display()
 
         # show the main window
         self.show()
@@ -226,11 +206,6 @@ class AoE2GameOverlay(RTSGameMatchDataOverlay):
         if self.selected_panel == PanelID.CONFIG:
             self.selected_panel = PanelID.BUILD_ORDER
         elif self.selected_panel == PanelID.BUILD_ORDER:
-            if is_valid_fetch_match_data(self.settings.fetch_match_data):
-                self.selected_panel = PanelID.MATCH_DATA
-            else:
-                self.selected_panel = PanelID.CONFIG
-        elif self.selected_panel == PanelID.MATCH_DATA:
             self.selected_panel = PanelID.CONFIG
 
         if self.selected_panel == PanelID.CONFIG:
@@ -316,10 +291,6 @@ class AoE2GameOverlay(RTSGameMatchDataOverlay):
         self.build_order_search.show()
         self.build_order_selection.show()
 
-        self.username_title.show()
-        self.username_search.show()
-        self.username_selection.show()
-
         # configuration buttons
         layout = self.settings.layout
         border_size = layout.border_size
@@ -372,24 +343,14 @@ class AoE2GameOverlay(RTSGameMatchDataOverlay):
 
         self.build_order_selection.update_size_position(init_y=next_y)
 
-        # username selection
-        layout_configuration = layout.configuration
-        next_x = layout_configuration.search_spacing + max(
+        next_x = max(
             widget_x_end(self.civilization_select), widget_x_end(self.build_order_search),
             self.build_order_selection.x() + self.build_order_selection.row_max_width)
 
-        self.username_title.move(next_x, self.build_order_title.y())
-        self.username_search.move(next_x, self.build_order_search.y())
-        self.username_selection.update_size_position(init_x=next_x, init_y=next_y)
-
-        max_x = max(widget_x_end(self.next_panel_button),
-                    widget_x_end(self.username_title), widget_x_end(self.username_search),
-                    self.username_selection.x() + self.username_selection.row_max_width)
+        max_x = widget_x_end(self.next_panel_button)
 
         max_y = max(widget_y_end(self.build_order_search),
-                    self.build_order_selection.y() + self.build_order_selection.row_total_height,
-                    widget_y_end(self.username_search),
-                    self.username_selection.y() + self.username_selection.row_total_height)
+                    self.build_order_selection.y() + self.build_order_selection.row_total_height)
 
         # resize main window
         self.resize(max_x + border_size, max_y + border_size)
@@ -561,321 +522,6 @@ class AoE2GameOverlay(RTSGameMatchDataOverlay):
         # position update to stay with the same upper right corner position
         self.update_position()
 
-    def fetch_game_match_data(self):
-        """Fetch the game match data"""
-        if self.selected_username is not None:  # only available if valid username
-            # new tread call can be launched
-            if (not self.match_data_thread_started) or (len(self.store_match_data) >= 1):
-
-                # update valid new match found if last url calls are done
-                if len(self.store_match_data) >= 1:
-                    if self.store_match_data[0] is not None:
-                        if self.store_match_data[0].match_id is not None:
-                            self.match_data = self.store_match_data[0]
-                        elif self.match_data is None:
-                            self.match_data_warnings = self.store_match_data[0].warnings
-                    self.store_match_data.clear()
-
-                # launch new thread search
-                if is_valid_fetch_match_data(self.settings.fetch_match_data):
-                    if self.match_data is None:
-                        self.match_data_thread_id = get_match_data_threading(
-                            self.settings.fetch_match_data, self.store_match_data, stop_event=self.match_data_stop_flag,
-                            search_input=self.selected_username, timeout=self.settings.url_timeout)
-                    else:
-                        self.match_data_thread_id = get_match_data_threading(
-                            self.settings.fetch_match_data, self.store_match_data, stop_event=self.match_data_stop_flag,
-                            search_input=self.selected_username, timeout=self.settings.url_timeout,
-                            last_match_id=self.match_data.match_id, last_data_found=self.match_data.all_data_found)
-                else:
-                    self.match_data_thread_id = None
-
-                self.match_data_thread_started = True
-
-    def update_match_data_display(self):
-        """Display match data panel"""
-        self.match_data_display.clear()
-
-        if self.selected_username is None:
-            self.match_data_display.add_row_from_picture_line(
-                parent=self, line='No username provided to find match data.')
-
-        elif self.match_data_thread_id is None:  # match data search not started
-            self.match_data_display.add_row_from_picture_line(
-                parent=self, line='Match data search not yet started.')
-
-        elif self.match_data is None:  # user match data not found
-            self.match_data_display.add_row_from_picture_line(
-                parent=self, line=f'No match found (yet) for \'{self.selected_username}\'.')
-            for warning_comment in self.match_data_warnings:
-                self.match_data_display.add_row_from_picture_line(parent=self, line=warning_comment)
-
-        else:  # valid match available
-            # check if some columns must be shown (available for one player or more)
-            country_show = False  # check if country flags must be shown
-            single_elo_show = False  # check if single ELO must be shown
-            win_loss_show = False  # check if wins and losses must be shown
-            for cur_player in self.match_data.players:
-                if cur_player.country is not None:
-                    country_show = True
-                if cur_player.elo_solo is not None:
-                    single_elo_show = True
-                if cur_player.win_rate is not None:
-                    win_loss_show = True
-
-            # color for the ELO
-            layout_match_data = self.settings.layout.match_data  # settings of the match data layout
-            color_elo = layout_match_data.color_elo if win_loss_show else layout_match_data.color_elo_no_win_loss
-            color_elo_solo = layout_match_data.color_elo_solo if win_loss_show else \
-                layout_match_data.color_elo_solo_no_win_loss
-
-            max_length = layout_match_data.match_data_max_length  # max length of the data to display
-
-            # separation spaces between elements
-            separation = '@'
-            for i in range(layout_match_data.resource_spacing):
-                separation += ' '
-            separation += '@'
-
-            # describe the title row with its settings
-            title_line = ''
-            title_labels_settings = []
-
-            # player color
-            title_line += ' '
-            title_labels_settings.append(None)
-
-            # civilization icon
-            title_line += separation + ' '
-            title_labels_settings.append(None)
-            title_labels_settings.append(None)
-
-            # player name column, used to display the map name
-            if self.match_data.map_name is not None:
-                title_line += separation + cut_name_length(self.match_data.map_name, max_length)
-            else:
-                title_line += separation + cut_name_length('Unknown map', max_length)
-            title_labels_settings.append(None)
-            title_labels_settings.append(
-                QLabelSettings(text_color=layout_match_data.color_map, text_bold=True, text_alignment='left'))
-
-            # single ELO show
-            if single_elo_show:
-                title_line += separation + 'Solo'
-                title_labels_settings.append(None)
-                title_labels_settings.append(QLabelSettings(text_color=color_elo_solo, text_bold=True,
-                                                            text_alignment='center'))
-
-            # game type ELO
-            if single_elo_show:
-                title_line += separation + 'Team'
-            else:
-                title_line += separation + 'Elo'
-            title_labels_settings.append(None)
-            title_labels_settings.append(
-                QLabelSettings(text_color=color_elo, text_bold=True, text_alignment='center'))
-
-            # player rank
-            title_line += separation + 'Rank'
-            title_labels_settings.append(None)
-            title_labels_settings.append(
-                QLabelSettings(text_color=layout_match_data.color_rank, text_bold=True, text_alignment='center'))
-
-            # player win rate
-            if win_loss_show:
-                title_line += separation + 'Win%'
-                title_labels_settings.append(None)
-                title_labels_settings.append(
-                    QLabelSettings(text_color=layout_match_data.color_win_rate, text_bold=True,
-                                   text_alignment='center'))
-
-            # player wins
-            if win_loss_show:
-                title_line += separation + 'Win'
-                title_labels_settings.append(None)
-                title_labels_settings.append(
-                    QLabelSettings(text_color=layout_match_data.color_wins, text_bold=True, text_alignment='center'))
-
-            # player losses
-            if win_loss_show:
-                title_line += separation + 'Loss'
-                title_labels_settings.append(None)
-                title_labels_settings.append(
-                    QLabelSettings(text_color=layout_match_data.color_losses, text_bold=True, text_alignment='center'))
-
-            # next panel (+ optional country flag)
-            title_line += separation + ' '
-            title_labels_settings.append(None)
-            title_labels_settings.append(None)
-
-            # display title line
-            self.match_data_display.add_row_from_picture_line(parent=self, line=title_line,
-                                                              labels_settings=title_labels_settings)
-
-            # loop on the players
-            for cur_player in self.match_data.players:
-                player_line = ''
-                player_labels_settings = []
-
-                # player color
-                color_str = '?'  # assuming unknown color string
-                color = self.settings.layout.color_default  # use default color
-
-                color_id = cur_player.color
-                if (color_id is not None) and isinstance(color_id, int):
-                    color_str = str(color_id)
-
-                    if color_id == 1:
-                        color = layout_match_data.color_player_1
-                    elif color_id == 2:
-                        color = layout_match_data.color_player_2
-                    elif color_id == 3:
-                        color = layout_match_data.color_player_3
-                    elif color_id == 4:
-                        color = layout_match_data.color_player_4
-                    elif color_id == 5:
-                        color = layout_match_data.color_player_5
-                    elif color_id == 6:
-                        color = layout_match_data.color_player_6
-                    elif color_id == 7:
-                        color = layout_match_data.color_player_7
-                    elif color_id == 8:
-                        color = layout_match_data.color_player_8
-                    else:
-                        print(f'Unknown color ID {color_id}.')
-
-                if not isinstance(color, list):
-                    print(f'Invalid player color {color}, using the default one.')
-                    color = self.settings.layout.color_default
-
-                player_line += color_str
-                player_labels_settings.append(QLabelSettings(text_alignment='center', text_color=color, text_bold=True))
-
-                # civilization icon
-                if cur_player.civ is not None:
-                    player_line += separation + f'civilization/{aoe2_civilization_icon[cur_player.civ][1]}' if (
-                            cur_player.civ in aoe2_civilization_icon) else cur_player.civ
-                else:
-                    player_line += separation + '?'
-                player_labels_settings.append(None)
-                player_labels_settings.append(QLabelSettings(text_alignment='center'))
-
-                # player name
-                if cur_player.name is not None:
-                    player_line += separation + cut_name_length(cur_player.name, max_length)
-                else:
-                    player_line += separation + 'Unknown'
-                player_labels_settings.append(None)
-                player_labels_settings.append(
-                    QLabelSettings(text_color=layout_match_data.color_player_name, text_alignment='left'))
-
-                # single ELO show
-                if single_elo_show:
-                    player_labels_settings.append(None)
-                    if cur_player.elo_solo is not None:
-                        player_line += separation + str(cur_player.elo_solo)
-                        player_labels_settings.append(
-                            QLabelSettings(text_color=color_elo_solo, text_alignment='right'))
-                    else:
-                        player_line += separation + '-'
-                        player_labels_settings.append(
-                            QLabelSettings(text_color=color_elo_solo, text_alignment='center'))
-
-                # game type ELO
-                if cur_player.elo is not None:
-                    player_line += separation + str(cur_player.elo)
-                else:
-                    player_line += separation + '-'
-                player_labels_settings.append(None)
-                player_labels_settings.append(
-                    QLabelSettings(text_color=color_elo, text_alignment='right'))
-
-                # player rank
-                if cur_player.rank is not None:
-                    player_line += separation + '#' + str(cur_player.rank)
-                else:
-                    player_line += separation + '-'
-                player_labels_settings.append(None)
-                player_labels_settings.append(
-                    QLabelSettings(text_color=layout_match_data.color_rank, text_alignment='right'))
-
-                # player win rate
-                if win_loss_show:
-                    if cur_player.win_rate is not None:
-                        player_line += separation + str(cur_player.win_rate) + '%'
-                    else:
-                        player_line += separation + '-'
-                    player_labels_settings.append(None)
-                    player_labels_settings.append(
-                        QLabelSettings(text_color=layout_match_data.color_win_rate, text_alignment='right'))
-
-                # player wins
-                if win_loss_show:
-                    if cur_player.wins is not None:
-                        player_line += separation + str(cur_player.wins)
-                    else:
-                        player_line += separation + '-'
-                    player_labels_settings.append(None)
-                    player_labels_settings.append(
-                        QLabelSettings(text_color=layout_match_data.color_wins, text_alignment='right'))
-
-                # player losses
-                if win_loss_show:
-                    if cur_player.losses is not None:
-                        player_line += separation + str(cur_player.losses)
-                    else:
-                        player_line += separation + '-'
-                    player_labels_settings.append(None)
-                    player_labels_settings.append(
-                        QLabelSettings(text_color=layout_match_data.color_losses, text_alignment='right'))
-
-                # country flag
-                player_labels_settings.append(None)
-                if country_show:
-                    flag_name = cur_player.country if (cur_player.country is not None) else 'unknown'
-                    country_flag_image = os.path.join(self.directory_common_pictures, 'national_flag',
-                                                      f'{flag_name.lower()}.png')
-                    player_line += separation + f'national_flag/{flag_name.lower()}.png' if os.path.isfile(
-                        country_flag_image) else flag_name
-                    player_labels_settings.append(QLabelSettings(image_width=layout_match_data.flag_width,
-                                                                 image_height=layout_match_data.flag_height,
-                                                                 text_alignment='center'))
-                else:
-                    player_line += separation
-                    for i in range(layout_match_data.flag_space_no_country):
-                        player_line += ' '
-                    player_labels_settings.append(None)
-
-                # display player line
-                self.match_data_display.add_row_from_picture_line(parent=self, line=player_line,
-                                                                  labels_settings=player_labels_settings)
-
-        self.game_match_data_layout()  # update layout
-
-    def game_match_data_layout(self):
-        """Layout of the game match panel"""
-        self.match_data_display.show()
-        self.next_panel_button.show()
-
-        # size and position
-        self.match_data_display.update_size_position(adapt_to_columns=True)
-
-        # resize of the full window
-        border_size = self.match_data_display.border_size
-
-        width = 2 * border_size + self.match_data_display.row_max_width
-        if self.match_data is None:  # increase size for the next frame button
-            width += self.settings.layout.horizontal_spacing + self.settings.layout.action_button_size
-
-        self.resize(width, 2 * border_size + self.match_data_display.row_total_height)
-
-        # next panel on the top right corner
-        self.next_panel_button.move(self.width() - border_size - self.next_panel_button.width(), border_size)
-        self.next_panel_button.raise_()  # raise to the top of the parent widget's stack
-
-        # update position (in case the size changed)
-        self.update_position()
-
     def timer_mouse_keyboard_call(self):
         """Function called on a timer (related to mouse and keyboard inputs)"""
         super().timer_mouse_keyboard_call()
@@ -900,24 +546,12 @@ class AoE2GameOverlay(RTSGameMatchDataOverlay):
                             tooltip, self.x() + label_x, self.y() + label_y,
                             self.settings.layout.build_order.tooltip_timeout)
 
-    def timer_match_data_call(self):
-        """Function called on a timer (related to match data)"""
-        if not self.stop_application:
-            self.fetch_game_match_data()
-
-            if self.selected_panel == PanelID.MATCH_DATA:
-                self.update_match_data_display()  # layout updated in function
-
     def enter_key_actions(self):
         """Actions performed when pressing the Enter key"""
         if self.selected_panel == PanelID.CONFIG:
             if self.build_order_search.hasFocus():
                 self.select_build_order(key_condition={
                     'civilization': self.civilization_combo_ids[self.civilization_select.currentIndex()]})
-
-            if self.username_search.hasFocus():
-                self.select_username()  # update username
-                self.fetch_game_match_data()  # launch potential new game search
 
             self.config_panel_layout()  # update layout